# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ProjectTnW - Multi-tenant School Transportation SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - in: query
          name: routeId
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"

    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{studentId}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── PARENTS ───
  /parents:
    get:
      operationId: listParents
      tags: [parents]
      summary: List parents
      responses:
        "200":
          description: Parents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Parent"

    post:
      operationId: createParent
      tags: [parents]
      summary: Create a parent record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parent"
  /parents/{parentId}/students:
    get:
      operationId: listParentStudents
      tags: [parents]
      summary: Get students linked to a parent
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"

    post:
      operationId: linkStudentToParent
      tags: [parents]
      summary: Link a student to a parent
      parameters:
        - in: path
          name: parentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkStudentBody"
      responses:
        "201":
          description: Linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── TRIPS ───
...
  /admin/tenants:
    get:
      operationId: adminListTenants
      tags: [super-admin]
      summary: List all tenants (super-admin)
      responses:
        "200":
          description: Tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"

    post:
      operationId: adminCreateTenant
      tags: [super-admin]
      summary: Create a tenant (super-admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateTenantBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /admin/tenants/{tenantId}:
    get:
      operationId: adminGetTenant
      tags: [super-admin]
      summary: Get a tenant (super-admin)
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

    patch:
      operationId: adminUpdateTenant
      tags: [super-admin]
      summary: Update tenant plan/status (super-admin)
      parameters:
        - in: path
          name: tenantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateTenantBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── AUTH ───
...
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        isDemo:
          type: ["boolean", "null"]
        plan:
          type: ["string", "null"]
        billingStatus:
          type: ["string", "null"]
        trialEndsAt:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - slug
...
    Student:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        schoolName:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        homeStopId:
          type: ["string", "null"]
        boardingTime:
          type: ["string", "null"]
        alightingTime:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        specialNeeds:
          type: ["string", "null"]
      required:
        - id
        - tenantId
        - firstName
        - lastName
    StudentDetail:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        schoolName:
          type: ["string", "null"]
        grade:
          type: ["integer", "null"]
        homeStopId:
          type: ["string", "null"]
        boardingTime:
          type: ["string", "null"]
        alightingTime:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        specialNeeds:
          type: ["string", "null"]
        parents:
          type: array
          items:
            $ref: "#/components/schemas/Parent"
      required:
        - id
        - tenantId
        - firstName
        - lastName
        - parents
    CreateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        schoolName:
          type: string
        grade:
          type: integer
        homeStopId:
          type: string
        boardingTime:
          type: string
        alightingTime:
          type: string
        emergencyContact:
          type: string
        specialNeeds:
          type: string
      required:
        - firstName
        - lastName
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        schoolName:
          type: string
        grade:
          type: integer
        homeStopId:
          type: ["string", "null"]
        boardingTime:
          type: string
        alightingTime:
          type: string
        emergencyContact:
          type: string
        specialNeeds:
          type: string

    # ─── PARENTS ───
    Parent:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        userId:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
      required:
        - id
        - tenantId
    CreateParentBody:
      type: object
      properties:
        userId:
          type: string
        phone:
          type: string
    LinkStudentBody:
      type: object
      properties:
        studentId:
          type: string
      required:
        - studentId

    # ─── TRIPS ───
...
    AdminCreateTenantBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
        isDemo:
          type: boolean
        adminEmail:
          type: string
        adminPassword:
          type: string
        adminFullName:
          type: string
      required:
        - name
        - slug
    AdminUpdateTenantBody:
      type: object
      properties:
        name:
          type: string
        plan:
          type: string
        billingStatus:
          type: string
        isDemo:
          type: boolean
        trialEndsAt:
          type: string
...